reorgenized react utils
